# specs/002-roles-permissions-system/openapi.yaml
openapi: 3.0.3
info:
  title: WasteFlow Roles & Permissions API
  description: |
    Comprehensive roles and permissions management system for WasteFlow application.

    This API supports:
    - Multi-tenant role-based access control (RBAC)
    - Attribute-based access control (ABAC) policies
    - Consultant multi-tenant context switching
    - Temporary permission grants with auto-expiration
    - Comprehensive audit logging for regulatory compliance
    - Custom role creation for enterprise clients
    - Permission request workflows

    ## Authentication
    All endpoints require JWT Bearer authentication obtained via SPID/CIE login.
    The JWT contains user identity, tenant context, and SPID fiscal code for audit traceability.

    ## Multi-Tenancy
    - Schema-per-tenant architecture with Row-Level Security (RLS)
    - Tenant context extracted from JWT claims
    - Consultants can switch tenant context via dedicated endpoints
    - All operations scoped to current tenant unless explicitly cross-tenant (consultants only)

    ## Rate Limiting
    - Standard endpoints: 100 requests/minute per user
    - Permission check endpoints: 1000 requests/minute per user (high-frequency caching)
    - Audit log queries: 20 requests/minute per user (expensive operations)
    - Bulk operations: 10 requests/minute per user

    ## Compliance
    - All audit logs retained for 10+ years per D.Lgs. 152/2006
    - Immutable audit trail with cryptographic integrity
...

  version: 1.0.0
...
servers:
...
security:
  - BearerAuth: []
paths:
...
  /consultant/tenants:
    get:
      tags: [Consultant Context]
      summary: List all tenant contexts for consultant
      description: |
        Returns all tenant associations for current consultant user.
        Each tenant shows consultant's role within that tenant.
        Regular (non-consultant) users receive empty array.
      operationId: listConsultantTenants
      responses:
        '200':
          description: Consultant tenants retrieved successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ConsultantTenantsResponse'
        '401':
          $ref: '#/components/responses/UnauthorizedError'
      x-rate-limit: 100/minute
  /consultant/switch-tenant:
    post:
      tags: [Consultant Context]
      summary: Switch active tenant context
      description: |
        Switches consultant's active tenant context.
        Issues new JWT with updated tenant claim.
        Frontend must reload to reflect new context.
        All subsequent API calls scoped to new tenant.
        Response time: <2 seconds at 95th percentile.
      operationId: switchTenantContext
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                tenantId:
                  type: string
                  format: uuid
                  description: Target tenant ID to switch to
              required: [tenantId]
      responses:
        '200':
          description: Tenant context switched successfully
          content:
            application/json:
              schema:
                type: object
                properties:
                  accessToken:
                    type: string
                    description: New JWT with updated tenant context
                  refreshToken:
                    type: string
                    description: New refresh token
                  expiresIn:
                    type: integer
                    description: Token expiration in seconds
                  tenant:
                    $ref: '#/components/schemas/TenantContextInfo'
        '400':
          $ref: '#/components/responses/ValidationError'
        '401':
          $ref: '#/components/responses/UnauthorizedError'
        '403':
          description: User not associated with target tenant
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
      x-rate-limit: 100/minute
  /consultant/dashboard:
    get:
      tags: [Consultant Context]
      summary: Get aggregated cross-tenant dashboard
      description: |
        Returns aggregated KPIs across all consultant's managed tenants:
        - Total pending FIRs by client
        - Upcoming MUD deadlines sorted by urgency
        - RENTRI sync failures color-coded by severity
        - Critical alerts requiring immediate attention

        Consultant-only endpoint - returns 403 for regular users.
      operationId: getConsultantAggregatedDashboard
      parameters:
        - name: includeClients
          in: query
          schema:
            type: array
            items:
              type: string
              format: uuid
          description: Filter to specific tenant IDs (default all)
        - name: urgencyLevel
          in: query
          schema:
            type: string
            enum: [critical, high, medium, low, all]
            default: all
          description: Filter alerts by urgency
      responses:
        '200':
          description: Aggregated dashboard retrieved successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ConsultantDashboardResponse'
        '401':
          $ref: '#/components/responses/UnauthorizedError'
        '403':
          description: Endpoint only available to consultant users
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
      x-rate-limit: 100/minute

  # ===========================
  # PERMISSION REQUESTS WORKFLOW
  # ===========================
...
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: |
        JWT token obtained via SPID/CIE authentication.
        Include in Authorization header: `Bearer <token>`

        JWT contains claims:
        - userId: User unique identifier
        - tenantId: Current tenant context
        - fiscalCode: SPID fiscal code for audit
        - roles: User's role IDs
        - permissions: Cached permission set
        - exp: Token expiration timestamp
...
  schemas:
...
    ConsultantTenantsResponse:
      type: object
      properties:
        consultantId:
          type: string
          format: uuid
        consultantFirm:
          type: string
          description: Consultant firm name
        tenants:
          type: array
          items:
            $ref: '#/components/schemas/TenantContextInfo'
        currentTenantId:
          type: string
          format: uuid
          description: Currently active tenant context
        totalClients:
          type: integer
      required: [consultantId, tenants, currentTenantId, totalClients]
    TenantContextInfo:
      type: object
      properties:
        tenantId:
          type: string
          format: uuid
        tenantName:
          type: string
        tenantType:
          type: string
          enum: [PRODUCER, TRANSPORTER, TREATMENT_FACILITY]
        role:
          type: string
          description: Consultant's role within this tenant
        roleId:
          type: string
          format: uuid
        permissions:
          type: array
          items:
            type: string
          description: Effective permissions in this tenant
        isActive:
          type: boolean
          description: Whether this is currently active context
        lastAccessedAt:
          type: string
          format: date-time
      required: [tenantId, tenantName, tenantType, role, roleId, isActive]
    ConsultantDashboardResponse:
      type: object
      properties:
        consultantId:
          type: string
          format: uuid
        totalClients:
          type: integer
        summary:
          type: object
          properties:
            totalPendingFirs:
              type: integer
            totalUpcomingDeadlines:
              type: integer
            criticalAlerts:
              type: integer
            syncFailures:
              type: integer
        clientAlerts:
          type: array
          items:
            type: object
            properties:
              tenantId:
                type: string
                format: uuid
              tenantName:
                type: string
              alertType:
                type: string
                enum: [PENDING_FIRS, MUD_DEADLINE, RENTRI_SYNC_FAILURE, COMPLIANCE_ISSUE]
              urgency:
                type: string
                enum: [CRITICAL, HIGH, MEDIUM, LOW]
              count:
                type: integer
              description:
                type: string
              actionUrl:
                type: string
                description: Deep link to resolve alert
              dueDate:
                type: string
                format: date-time
                nullable: true
          description: Alerts requiring attention sorted by urgency
        firsByClient:
          type: array
          items:
            type: object
            properties:
              tenantId:
                type: string
                format: uuid
              tenantName:
                type: string
              pendingCount:
                type: integer
              inProgressCount:
                type: integer
              completedThisMonth:
                type: integer
          description: FIR statistics per client
        upcomingDeadlines:
          type: array
          items:
            type: object
            properties:
              tenantId:
                type: string
                format: uuid
              tenantName:
                type: string
              deadlineType:
                type: string
              description:
                type: string
              dueDate:
                type: string
                format: date-time
              daysRemaining:
                type: integer
              urgency:
                type: string
                enum: [CRITICAL, HIGH, MEDIUM, LOW]
          description: Upcoming regulatory deadlines sorted by due date
        generatedAt:
          type: string
          format: date-time
      required: [consultantId, totalClients, summary, clientAlerts]

    # ===========================
    # PERMISSION REQUEST SCHEMAS
    # ===========================
...
    ErrorResponse:
      type: object
      properties:
        error:
          type: string
          description: Machine-readable error code
          example: PERMISSION_DENIED
        message:
          type: string
          description: Human-readable error message (Italian)
          example: Non hai il permesso necessario per eseguire questa azione
        statusCode:
          type: integer
          description: HTTP status code
        details:
          type: object
          description: Additional error context
          additionalProperties: true
        timestamp:
          type: string
          format: date-time
        path:
          type: string
          description: API endpoint that generated error
      required: [error, message, statusCode, timestamp]
    ValidationErrorDetail:
      type: object
      properties:
        field:
          type: string
          description: Field that failed validation
        message:
          type: string
          description: Validation error message
        value:
          description: Invalid value provided
        constraint:
          type: string
          description: Validation constraint violated
      required: [field, message]
  responses:
    UnauthorizedError:
      description: Authentication required
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
          example:
            error: UNAUTHORIZED
            message: Authentication required. Please login with SPID/CIE.
            statusCode: 401
            timestamp: 2025-10-31T14:30:00.000Z
            path: /v1/roles
...
    ValidationError:
      description: Request validation failed
      content:
        application/json:
          schema:
            allOf:
              - $ref: '#/components/schemas/ErrorResponse'
              - type: object
                properties:
                  validationErrors:
                    type: array
                    items:
                      $ref: '#/components/schemas/ValidationErrorDetail'
          example:
            error: VALIDATION_ERROR
            message: Request validation failed
            statusCode: 400
            timestamp: 2025-10-31T14:30:00.000Z
            path: /v1/roles
            validationErrors:
              - field: name
                message: Role name must be at least 3 characters
                value: ab
                constraint: minLength
              - field: permissions
                message: At least one permission is required
                constraint: minItems
...